告警清理

+++ services/core/attest/attest_service_active.c
@@ -99,14 +99,14 @@ int32_t ParseActiveResult(const char* jsonStr)
         ATTEST_LOG_ERROR("[ParseActiveResult] Invalid parameter");
         return ATTEST_ERR;
     }
-    uint64_t errorCode = GetObjectItemValueNumber(jsonStr, "errcode");
+    double errorCode = GetObjectItemValueNumber(jsonStr, "errcode");
     if (isnan(errorCode)) {
         ATTEST_LOG_ERROR("[ParseActiveResult] errorCode is nan.");
         return ATTEST_ERR;
     }
-    if (errorCode != ATTEST_OK) {
-        ATTEST_LOG_ERROR("[ParseActiveResult] -errorCode = %d.", -errorCode);
-        return -(errorCode);
+    if ((int32_t)errorCode != ATTEST_OK) {
+        ATTEST_LOG_ERROR("[ParseActiveResult] -errorCode = %d.", -(int32_t)(errorCode));
+        return -(int32_t)((errorCode));
     }
     return ATTEST_OK;
 }
